docker/dev/controller-requirements.txt
websockets==10.1
requests
pyyaml
